+++ CGES/src/renderer_phoneshader.cpp
@@ -9,8 +9,8 @@ namespace cges::renderer {
 void PhoneShader::ParallelDraw(const Camera& camera,
                                RenderBuffer& renderTarget,
                                const Scene& scene,
-                               const int loopMin,
-                               const int loopMax,
+                               const size_t loopMin,
+                               const size_t loopMax,
                                const glm::vec3& initialPos,
                                const glm::vec3& screenVerticalVec,
                                const glm::vec3& screenHorizontalVec) const {
